print-atm.c no longer performs special handling for FORE headers; these
 can no doubt be re-added at a later date.

 print-fr.c is effectively a no-op.

 print-llc.c has had the default_print_unaligned() call removed as
 tcpdump no longer defines this function, however the prototype is still
 present. Suggest we roll in a diff to use print_unknown_data().


Former-commit-id: 5a2aff39c8d94e985c9f4d68f014801ec9a6ce50

+++ contrib/tcpdump/token.h
@@ -1,6 +1,6 @@
 begin_unit|revision:0.9.5;language:C;cregit-version:0.0.1
 begin_comment
-comment|/* @(#) $Header: /tcpdump/master/tcpdump/token.h,v 1.3 2000/10/03 02:55:03 itojun Exp $ (LBL) */
+comment|/* @(#) $Header: /tcpdump/master/tcpdump/token.h,v 1.6 2002/12/11 07:14:12 guy Exp $ (LBL) */
 end_comment
 
 begin_comment
@@ -126,7 +126,7 @@ name|SEGMENT_COUNT
 parameter_list|(
 name|trp
 parameter_list|)
-value|((RIF_LENGTH(trp) - 2) / 2)
+value|((int)((RIF_LENGTH(trp) - 2) / 2))
 end_define
 
 begin_struct
